pyiceberg/requirements.txt
pyiceberg[pyarrow]
pyarrow
pandas
pyyaml